gcp_deployment/cloud_functions/requirements.txt
requests>=2.31.0
google-cloud-bigquery>=3.11.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
google-cloud-storage==2.13.0
pandas==2.1.4
pytz==2023.3
